- [Features](#features) - [Installation](#installation) - [Usage](#usage) - [Contributing](#contributing) - [License](#license)
Provide an overview of the project and its purpose.

List the main features and functionalities of the project.

Provide instructions on how to install and set up the project.

Explain how to use the project and provide examples if applicable.

Specify guidelines for contributing to the project, if applicable.

Indicate the license under which the project is distributed.